Moved the Disco animation from the ConfirmRobotName component to the WelcomeModal
