+++ app/src/organisms/OnDeviceDisplay/NameRobot/__tests__/ConfirmRobotName.test.tsx
@@ -5,10 +5,6 @@ import { renderWithProviders } from '@opentrons/components'
 
 import { i18n } from '../../../../i18n'
 import { ConfirmRobotName } from '../ConfirmRobotName'
-import { useCreateLiveCommandMutation } from '@opentrons/react-api-client'
-
-import type { SetStatusBarCreateCommand } from '@opentrons/shared-data/protocol/types/schemaV7/command/incidental'
-
 const mockPush = jest.fn()
 
 jest.mock('react-router-dom', () => {
@@ -18,13 +14,6 @@ jest.mock('react-router-dom', () => {
     useHistory: () => ({ push: mockPush } as any),
   }
 })
-
-jest.mock('@opentrons/react-api-client')
-jest.mock('@opentrons/shared-data/protocol/types/schemaV7/command/incidental')
-
-const mockUseCreateLiveCommandMutation = useCreateLiveCommandMutation as jest.MockedFunction<
-  typeof useCreateLiveCommandMutation
->
 
 const render = (props: React.ComponentProps<typeof ConfirmRobotName>) => {
   return renderWithProviders(
@@ -39,16 +28,10 @@ const render = (props: React.ComponentProps<typeof ConfirmRobotName>) => {
 
 describe('ConfirmRobotName', () => {
   let props: React.ComponentProps<typeof ConfirmRobotName>
-  let mockCreateLiveCommand = jest.fn()
   beforeEach(() => {
-    mockCreateLiveCommand = jest.fn()
-    mockCreateLiveCommand.mockResolvedValue(null)
     props = {
       robotName: 'otie',
     }
-    mockUseCreateLiveCommandMutation.mockReturnValue({
-      createLiveCommand: mockCreateLiveCommand,
-    } as any)
   })
 
   it('should render text, an image and a button', () => {
@@ -61,16 +44,7 @@ describe('ConfirmRobotName', () => {
   it('when tapping a button, call a mock function', () => {
     const [{ getByText }] = render(props)
     const button = getByText('Finish setup')
-    const animationCommand: SetStatusBarCreateCommand = {
-      commandType: 'setStatusBar',
-      params: { animation: 'disco' },
-    }
     button.click()
     expect(mockPush).toBeCalledWith('/dashboard')
-    expect(mockUseCreateLiveCommandMutation).toBeCalledWith()
-    expect(mockCreateLiveCommand).toBeCalledWith({
-      command: animationCommand,
-      waitUntilComplete: false,
-    })
   })
 })
