+++ app/src/pages/OnDeviceDisplay/RobotDashboard/__tests__/WelcomeModal.test.tsx
@@ -5,8 +5,17 @@ import { renderWithProviders } from '@opentrons/components'
 import { i18n } from '../../../../i18n'
 import { updateConfigValue } from '../../../../redux/config'
 import { WelcomeModal } from '../WelcomeModal'
+import { useCreateLiveCommandMutation } from '@opentrons/react-api-client'
+
+import type { SetStatusBarCreateCommand } from '@opentrons/shared-data/protocol/types/schemaV7/command/incidental'
 
 jest.mock('../../../../redux/config')
+jest.mock('@opentrons/react-api-client')
+jest.mock('@opentrons/shared-data/protocol/types/schemaV7/command/incidental')
+
+const mockUseCreateLiveCommandMutation = useCreateLiveCommandMutation as jest.MockedFunction<
+  typeof useCreateLiveCommandMutation
+>
 
 const mockFunc = jest.fn()
 const WELCOME_MODAL_IMAGE_NAME = 'welcome_dashboard_modal.png'
@@ -23,16 +32,26 @@ const render = (props: React.ComponentProps<typeof WelcomeModal>) => {
 
 describe('WelcomeModal', () => {
   let props: React.ComponentProps<typeof WelcomeModal>
+  let mockCreateLiveCommand = jest.fn()
 
   beforeEach(() => {
+    mockCreateLiveCommand = jest.fn()
+    mockCreateLiveCommand.mockResolvedValue(null)
     props = {
       setShowWelcomeModal: mockFunc,
     }
+    mockUseCreateLiveCommandMutation.mockReturnValue({
+      createLiveCommand: mockCreateLiveCommand,
+    } as any)
   })
 
   it('should render text and button', () => {
     const [{ getByText, getByRole }] = render(props)
     const image = getByRole('img')
+    const animationCommand: SetStatusBarCreateCommand = {
+      commandType: 'setStatusBar',
+      params: { animation: 'disco' },
+    }
 
     expect(image.getAttribute('src')).toEqual(WELCOME_MODAL_IMAGE_NAME)
     getByText('Welcome to your dashboard!')
@@ -40,6 +59,11 @@ describe('WelcomeModal', () => {
       'A place to run protocols, manage your instruments, and view robot status.'
     )
     getByText('Got it')
+    expect(mockUseCreateLiveCommandMutation).toBeCalledWith()
+    expect(mockCreateLiveCommand).toBeCalledWith({
+      command: animationCommand,
+      waitUntilComplete: false,
+    })
   })
 
   it('should call a mock function when tapping got it button', () => {
